refactor: create ShaderBatchBasedOverlay base class for common usage
