+++ src/modules/sbstudio/plugin/overlays/base.py
@@ -6,11 +6,15 @@ from __future__ import annotations
 
 from abc import ABCMeta
 from collections.abc import Callable
-from typing import ClassVar
+from typing import TYPE_CHECKING, ClassVar
 
 import bpy
 import gpu
+import gpu.state
 from bpy.types import SpaceView3D
+
+if TYPE_CHECKING:
+    from gpu.types import GPUBatch
 
 __all__ = ("Overlay",)
 
@@ -119,3 +123,45 @@ class ShaderOverlay(Overlay):
 
     def dispose(self) -> None:
         self._shader = None
+
+
+class ShaderBatchBasedOverlay(ShaderOverlay):
+    """SharedOverlay subclass that uses shader batches for redrawing."""
+
+    _shader_batches: list[GPUBatch] | None = None
+
+    def dispose(self) -> None:
+        super().dispose()
+        self._shader_batches = None
+
+    def _create_shader_batches(self) -> list[GPUBatch]:
+        """Creates shader batches dynamically.
+
+        Function should be overriden by child classes with properly
+        initializing shader batches for drawing on 3D Views.
+        """
+        return []
+
+    def _draw_shader_batches(
+        self, point_size: int | None = None, line_width: int | None = None
+    ):
+        """Draws shader batches to 3D Views.
+
+        Function should be typically called from the `draw_3d()`
+        method of child classes."""
+
+        assert self._shader is not None
+
+        gpu.state.blend_set("ALPHA")
+
+        if self._shader_batches is None:
+            self._shader_batches = self._create_shader_batches()
+
+        if self._shader_batches:
+            self._shader.bind()
+            if line_width is not None:
+                gpu.state.line_width_set(line_width)
+            if point_size is not None:
+                gpu.state.point_size_set(point_size)
+            for batch in self._shader_batches:
+                batch.draw(self._shader)
